# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "FeatureFlagOut": {
        "properties": {
          "default": {
            "title": "Default",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "enabled",
          "default",
          "description"
        ],
        "title": "FeatureFlagOut",
        "type": "object"
      },
      "FeatureFlagUpdate": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "FeatureFlagUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InventoryTransactionOut": {
        "properties": {
          "balance_after": {
            "title": "Balance After",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "product": {
            "$ref": "#/components/schemas/ProductSummary"
          },
          "qty_change": {
            "title": "Qty Change",
            "type": "integer"
          },
          "txn_type": {
            "$ref": "#/components/schemas/InventoryTxnType"
          }
        },
        "required": [
          "id",
          "product",
          "order_id",
          "txn_type",
          "qty_change",
          "balance_after",
          "note",
          "created_by",
          "created_at"
        ],
        "title": "InventoryTransactionOut",
        "type": "object"
      },
...
      "JobStatus": {
        "enum": [
          "PENDING",
          "RUNNING",
          "RETRY",
          "DONE",
          "DEAD",
          "EXPIRED"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "Page_JobOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[JobOut]",
        "type": "object"
      },
...
      "Page_ProductOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ProductOut]",
        "type": "object"
      },
...
      "ProductCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "opening_stock": {
            "default": 0,
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Opening Stock",
            "type": "integer"
          },
          "reorder_level": {
            "default": 10,
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Reorder Level",
            "type": "integer"
          },
          "sku": {
            "maxLength": 50,
            "minLength": 2,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Sku",
            "type": "string"
          },
          "unit_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Unit Price"
          }
        },
        "required": [
          "sku",
          "name",
          "unit_price"
        ],
        "title": "ProductCreate",
        "type": "object"
      },
      "ProductOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_low_stock": {
            "title": "Is Low Stock",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reorder_level": {
            "title": "Reorder Level",
            "type": "integer"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "stock_qty": {
            "title": "Stock Qty",
            "type": "integer"
          },
          "unit_price": {
            "title": "Unit Price",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "name",
          "description",
          "unit_price",
          "stock_qty",
          "reorder_level",
          "is_active",
          "is_low_stock",
          "created_at",
          "updated_at"
        ],
        "title": "ProductOut",
        "type": "object"
      },
...
      "ProductUpdate": {
        "additionalProperties": false,
        "description": "Stock is deliberately not editable here; use stock adjustments so the ledger stays complete.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "reorder_level": {
            "anyOf": [
              {
                "maximum": 1000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reorder Level"
          },
          "sku": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 2,
                "pattern": "^[A-Za-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          },
          "unit_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Price"
          }
        },
        "title": "ProductUpdate",
        "type": "object"
      },
...
      "StockAdjustmentCreate": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "quantity": {
            "description": "Positive to add stock; negative allowed only for ADJUSTMENT",
            "title": "Quantity",
            "type": "integer"
          },
          "txn_type": {
            "$ref": "#/components/schemas/InventoryTxnType",
            "default": "RESTOCK"
          }
        },
        "required": [
          "quantity"
        ],
        "title": "StockAdjustmentCreate",
        "type": "object"
      },
      "SystemStatusOut": {
        "properties": {
          "database": {
            "$ref": "#/components/schemas/DependencyStatus"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "jobs": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Jobs",
            "type": "object"
          },
          "migrations": {
            "$ref": "#/components/schemas/DependencyStatus"
          },
          "oldest_due_job_age_seconds": {
            "title": "Oldest Due Job Age Seconds",
            "type": "integer"
          },
          "redis": {
            "$ref": "#/components/schemas/DependencyStatus"
          },
          "scheduled_tasks": {
            "items": {
              "$ref": "#/components/schemas/ScheduledTaskOut"
            },
            "title": "Scheduled Tasks",
            "type": "array"
          },
          "uptime_seconds": {
            "title": "Uptime Seconds",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "workers": {
            "items": {
              "$ref": "#/components/schemas/WorkerOut"
            },
            "title": "Workers",
            "type": "array"
          }
        },
        "required": [
          "version",
          "environment",
          "uptime_seconds",
          "database",
          "redis",
          "migrations",
          "workers",
          "jobs",
          "oldest_due_job_age_seconds",
          "scheduled_tasks"
        ],
        "title": "SystemStatusOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "The access_token returned by /auth/otp/verify",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Products, customers, sales orders, inventory and a manager approval workflow.",
    "title": "Sales & Inventory Management System",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/feature-flags": {
      "get": {
        "operationId": "list_flags_api_v1_feature_flags_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeatureFlagOut"
                  },
                  "title": "Response List Flags Api V1 Feature Flags Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Feature flags (any signed-in user)",
        "tags": [
          "System"
        ]
      }
    },
    "/api/v1/feature-flags/{name}": {
      "put": {
        "operationId": "set_flag_api_v1_feature_flags__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureFlagUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureFlagOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Switch a feature on or off (audited)",
        "tags": [
          "System"
        ]
      }
    },
...
    "/api/v1/products": {
      "get": {
        "operationId": "list_products_api_v1_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "is_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "in": "query",
            "name": "low_stock",
            "required": false,
            "schema": {
              "default": false,
              "title": "Low Stock",
              "type": "boolean"
            }
          },
          {
            "description": "1-based page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-based page number",
              "maximum": 10000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ProductOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Products",
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "create_product_api_v1_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Product",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{product_id}": {
      "delete": {
        "operationId": "deactivate_product_api_v1_products__product_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Deactivate a product (soft delete)",
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "get_product_api_v1_products__product_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Product",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "update_product_api_v1_products__product_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Product",
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{product_id}/stock-adjustments": {
      "post": {
        "description": "Supports an `Idempotency-Key` header, so a retried restock is applied once.",
        "operationId": "adjust_stock_api_v1_products__product_id__stock_adjustments_post",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockAdjustmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryTransactionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Restock or manually adjust stock (recorded in the inventory ledger)",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/v1/system/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_system_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "job_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Type"
            }
          },
          {
            "description": "1-based page number",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-based page number",
              "maximum": 10000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Background jobs (outbox), newest first",
        "tags": [
          "System"
        ]
      }
    },
    "/api/v1/system/jobs/{job_id}/retry": {
      "post": {
        "operationId": "retry_job_api_v1_system_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry a dead-lettered job",
        "tags": [
          "System"
        ]
      }
    },
    "/api/v1/system/status": {
      "get": {
        "operationId": "system_status_api_v1_system_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatusOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Health of the API, database, cache and workers",
        "tags": [
          "System"
        ]
      }
    },
...
  }
}